bench/requirements.txt
# Benchmarking dependencies
psutil>=5.9.0
pyzmq>=25.0.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0

# Serialization dependencies
msgpack>=1.0.0

# Optional profiling tools
py-spy>=0.3.14  # For flamegraphs
memory-profiler>=0.61.0  # For memory profiling
